fix: разнесены model сервисы по модулям
